# swagger.yaml
openapi: 3.0.0
info:
  title: Teamitaka Practice API
  version: 1.0.0
  description: |
    Teamitaka 프로젝트 연습을 위한 API 명세서로서, 사용자 관리, 인증, 프로젝트 관리, 모집 공고, 프로필, 댓글, 스크랩, 검색, 대학 인증 등의 기능을 제공합니다.
    이 API는 Node.js와 Express.js를 기반으로 구현되었으며, Swagger UI를 통해 시각화됩니다.  
    주요 기능:
    - 사용자 등록 및 인증 (JWT 기반)
    - 관리자 기능 (관리자 계정 관리, 인증된 이메일 관리)
    - 모집 공고 생성/수정/삭제 및 댓글 관리
    - 프로젝트 생성/수정/삭제 및 멤버 관리
    - 프로필 생성/수정/삭제 및 이력서 조회
    - 스크랩 기능
    - 키워드 기반 검색
    - 대학 인증 (Univecert API 연동)
  contact:
    name: Teamitaka Support
    email: support@teamitaka.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:{port}
    description: 로컬 개발 서버
    variables:
      port:
        default: "3000"
        description: "서버 포트 (동적으로 설정 가능)"
  - url: https://api.teamitaka.com
    description: 프로덕션 서버
...
paths:
...
  /api/recruitment:
    get:
      tags:
        - Recruitment
      summary: 모집 공고 목록 조회
      description: 모든 모집 공고를 페이지네이션으로 반환합니다.
      operationId: getRecruitments
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 페이지 번호
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: 페이지당 항목 수
        - name: status
          in: query
          schema:
            type: string
            enum: ["OPEN", "CLOSED", "ALL"]
            default: "ALL"
          description: 모집 공고 상태 필터링
        - name: sort
          in: query
          schema:
            type: string
            enum: ["created_at_asc", "created_at_desc", "title_asc", "title_desc"]
            default: "created_at_desc"
          description: 정렬 기준
      responses:
        '200':
          description: 모집 공고 목록 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  recruitments:
                    - recruitment_id: 1
                      title: "개발자 모집"
                      description: "프로젝트 팀원 모집"
                      status: "OPEN"
                      created_at: "2025-02-27T11:00:00.000Z"
                      updated_at: "2025-02-27T11:00:00.000Z"
                  total: 1
                  page: 1
                  limit: 10
                message: ""
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
    post:
      tags:
        - Recruitment
      summary: 모집 공고 생성
      description: 새로운 모집 공고를 생성합니다.
      operationId: createRecruitment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 100
                  example: "개발자 모집"
                  description: 모집 공고 제목 (3~100자)
                description:
                  type: string
                  minLength: 10
                  maxLength: 1000
                  example: "프로젝트 팀원 모집"
                  description: 모집 공고 설명 (10~1000자)
                status:
                  type: string
                  enum: ["OPEN", "CLOSED"]
                  default: "OPEN"
                  description: 모집 공고 상태
              required:
                - title
                - description
      responses:
        '201':
          description: 모집 공고 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  recruitment_id: 1
                  title: "개발자 모집"
                  description: "프로젝트 팀원 모집"
                  status: "OPEN"
                  created_at: "2025-02-27T11:00:00.000Z"
                  updated_at: "2025-02-27T11:00:00.000Z"
                message: "모집 공고 생성 완료"
        '400':
          description: 필수 필드 누락 또는 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "제목은 3~100자 이내여야 합니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/recruitment/{recruitment_id}:
    get:
      tags:
        - Recruitment
      summary: 특정 모집 공고 조회
      description: 특정 모집 공고의 상세 정보를 반환합니다.
      operationId: getRecruitmentById
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 조회할 모집 공고 ID
      responses:
        '200':
          description: 모집 공고 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  recruitment_id: 1
                  title: "개발자 모집"
                  description: "프로젝트 팀원 모집"
                  status: "OPEN"
                  created_at: "2025-02-27T11:00:00.000Z"
                  updated_at: "2025-02-27T11:00:00.000Z"
                message: ""
        '404':
          description: 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
    put:
      tags:
        - Recruitment
      summary: 모집 공고 수정
      description: 특정 모집 공고의 정보를 수정합니다.
      operationId: updateRecruitment
      security:
        - bearerAuth: []
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 수정할 모집 공고 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 100
                  example: "개발자 모집 수정"
                  description: 수정할 제목 (3~100자)
                description:
                  type: string
                  minLength: 10
                  maxLength: 1000
                  example: "프로젝트 팀원 모집 (수정됨)"
                  description: 수정할 설명 (10~1000자)
                status:
                  type: string
                  enum: ["OPEN", "CLOSED"]
                  description: 수정할 상태
              required:
                - title
                - description
      responses:
        '200':
          description: 모집 공고 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  recruitment_id: 1
                  title: "개발자 모집 수정"
                  description: "프로젝트 팀원 모집 (수정됨)"
                  status: "OPEN"
                  updated_at: "2025-02-27T11:00:00.000Z"
                message: "모집 공고 수정 완료"
        '400':
          description: 필수 필드 누락 또는 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "제목은 3~100자 이내여야 합니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components.schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고 수정 권한이 없습니다."
        '404':
          description: 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
    delete:
      tags:
        - Recruitment
      summary: 모집 공고 삭제
      description: 특정 모집 공고를 삭제합니다.
      operationId: deleteRecruitment
      security:
        - bearerAuth: []
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 삭제할 모집 공고 ID
      responses:
        '200':
          description: 모집 공고 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data: {}
                message: "모집 공고 삭제 완료"
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고 삭제 권한이 없습니다."
        '404':
          description: 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"

  # 프로젝트 관련 엔드포인트
...
  /api/scraps/recruitments:
    get:
      tags:
        - Scrap
      summary: 스크랩한 모집 공고 목록 조회
      description: 현재 사용자가 스크랩한 모집 공고 목록을 페이지네이션으로 반환합니다.
      operationId: getScrapedRecruitments
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 페이지 번호
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: 페이지당 항목 수
        - name: sort
          in: query
          schema:
            type: string
            enum: ["created_at_asc", "created_at_desc", "title_asc", "title_desc"]
            default: "created_at_desc"
          description: 정렬 기준
      responses:
        '200':
          description: 스크랩 목록 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  recruitments:
                    - recruitment_id: 1
                      user_id: 1
                      title: "개발자 모집"
                      description: "프로젝트 팀원 모집"
                      status: "OPEN"
                      scrap_count: 5
                      created_at: "2025-02-27T11:00:00.000Z"
                      updated_at: "2025-02-27T11:00:00.000Z"
                  total: 1
                  page: 1
                  limit: 10
                message: ""
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/recruitment/{recruitment_id}/scrap:
    put:
      tags:
        - Scrap
      summary: 모집 공고 스크랩 토글
      description: 특정 모집 공고를 스크랩하거나 스크랩을 취소합니다.
      operationId: toggleScrap
      security:
        - bearerAuth: []
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 스크랩할 모집 공고 ID
      responses:
        '200':
          description: 스크랩 상태 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  action: "스크랩 추가"
                  recruitment_id: 1
                  user_id: 1
                  scrap_count: 6
                message: "스크랩 상태 변경 완료"
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '404':
          description: 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"

  # 검색 관련 엔드포인트
...
  /api/univcert/send-otp:
    post:
      tags:
        - UnivCert
      summary: OTP 발송
      description: 지정된 대학 이메일로 인증용 OTP를 발송합니다 (Univecert API 연동). 요율 제한이 적용됩니다.
      operationId: sendOtp
      x-rate-limit:
        limit: 5
        period: 60
        description: 분당 5회로 요청 제한
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "student@korea.ac.kr"
                  description: 인증할 대학 이메일 주소
              required:
                - email
      responses:
        '200':
          description: OTP 발송 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  message: "OTP 발송 완료"
                message: ""
        '400':
          description: 유효하지 않은 이메일
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "지원되지 않는 대학 이메일입니다."
        '429':
          description: 요청 횟수 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "요청 횟수를 초과했습니다. 잠시 후 다시 시도하세요."
        '500':
          description: 서버 오류 또는 외부 API 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "외부 API 호출 중 오류 발생"
  /api/univcert/verify-otp:
    post:
      tags:
        - UnivCert
      summary: OTP 검증
      description: 발송된 OTP 코드를 검증하여 대학 인증을 완료합니다.
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "student@korea.ac.kr"
                  description: 인증할 대학 이메일 주소
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "123456"
                  description: 발송된 OTP 코드 (6자리 숫자)
              required:
                - email
                - code
      responses:
        '200':
          description: OTP 검증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  certified_email: "student@korea.ac.kr"
                  certified_date: "2025-02-27T11:00:00.000Z"
                message: "대학 인증 성공"
        '400':
          description: 잘못된 OTP 코드 또는 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "잘못된 OTP 코드입니다."
        '500':
          description: 서버 오류 또는 외부 API 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "외부 API 호출 중 오류 발생"

  # 서버 상태 확인 엔드포인트
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        data:
          type: object
          description: 반환 데이터
        message:
          type: string
          description: 추가 메시지
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: 요청 실패 여부
        message:
          type: string
          description: 오류 메시지
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 기반 인증 토큰 (Bearer 형식)
...